--- frontend/src/metabase/components/MetadataInfo/DimensionInfo/index.js
@@ -1,1 +1,0 @@
-export { default } from "./DimensionInfo";
