--- frontend/src/metabase/components/MetadataInfo/DimensionInfoPopover/index.js
@@ -1,1 +1,0 @@
-export { default } from "./DimensionInfoPopover";
