+++ frontend/src/metabase/components/MetadataInfo/FieldInfo/index.js
@@ -1,0 +1,1 @@
+export { default } from "./FieldInfo";
