+++ frontend/src/metabase/components/MetadataInfo/FieldInfoPopover/index.js
@@ -1,0 +1,1 @@
+export { default } from "./FieldInfoPopover";
