+++ frontend/src/metabase/query_builder/components/dataref/FieldPane.tsx
@@ -1,4 +1,4 @@
-import DimensionInfo from "metabase/components/MetadataInfo/DimensionInfo";
+import DimensionInfo from "metabase/components/MetadataInfo/FieldInfo";
 import SidebarContent from "metabase/query_builder/components/SidebarContent";
 import type Field from "metabase-lib/metadata/Field";
 import { PaneContent } from "./Pane.styled";
